requirements.txt
langchain==0.0.196
pinecone_client==2.2.2
altair<5
streamlit==1.14.0
streamlit_chat==0.0.2.2
openai
gTTS==2.3.2
streamlit_option_menu==0.3.6
streamlit_lottie==0.0.3
Requests==2.31.0